# app/Http/Requests/Admin/ReasonRequest.php
<?php

namespace App\Http\Requests\Admin;

use Illuminate\Foundation\Http\FormRequest;

class ReasonRequest extends FormRequest
{
    public function authorize(): bool { return $this->user()?->isAdmin() === true; }
    public function rules(): array { return ['reason' => ['required', 'string', 'min:3', 'max:2000']]; }
}