# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WellSharp Same-Origin JSON Contract
  version: 1.2.0
  description: |
    JSON endpoints used by the authenticated Laravel Blade/Alpine.js interfaces.
    This is not a bearer-token REST API: the application uses Laravel session
    authentication and CSRF protection. There is no routes/api.php in the project.
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: http://localhost
...
paths:
...
  /admin/certificates/{certificate}/knowledge-controls/{control}/revert:
    patch:
      tags: [Admin Knowledge controls]
      security: [{ laravelSession: [] }]
      operationId: revertKnowledgeControl
      summary: Revert one active Knowledge control without deleting its history
      parameters:
        - { name: certificate, in: path, required: true, schema: { type: string } }
        - { name: control, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '302': { description: Redirect back to Admin Certificate details. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    laravelSession:
      type: apiKey
      in: cookie
      name: laravel_session
      description: Laravel session cookie. State-changing browser calls also require the CSRF token.
...
  schemas:
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 2000 }
...
  responses:
...
    Forbidden: { description: Authorization failed. }
    NotFound: { description: Resource not found. }
...
    ValidationError:
      description: Laravel validation or business-rule failure.
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors: { type: object, additionalProperties: { type: array, items: { type: string } } }
...